Added the ability to use a local client patch.
For testing purposes you can save any custom client patch in your bot cache folder with the name "modscript" (should NOT be in gzip format). The client will then reparse the client every time on startup, using the classes from memory for the entire session. The classes are still saved to the client.jar file for debugging purposes.